removed debug output, split scripts
