+++ isaaq/Scheduler/__init__.py
@@ -1,1 +1,2 @@
+from isaaq.Scheduler.BaseQAPScheduler import *
 from isaaq.Scheduler.QAPSchedulers import *
